refactor(ansible): Major refactor to ansible plays
The separate cmd and base dirs are now combined to utilize common roles.
As of now, docker_setup role does not start the respective containers
due to differences in the folder structure and starting commands. I'm
not sure if this is even meaningful to write as there are future plans
to move away from docker-compose.