feat: add renew token script for periodic token
This commit adds renew_token.sh to ensure consul-template's auth token
does not ever expire until it is restarted. A helper script is required
as renewing tokens from cert auth requires the client cert and key,
which prevents the use of consul-template's intrinsic token renewal
feature.

The commit also uses the newest pkiCert templating in consul-template to
stop the re-issuing of non-expired certs on every reload of
consul-template.